verbs: Extend async_event function in all providers to get ibv_context
The mlx5 driver needs to get the context of the IB device in case an
IBV_EVENT_DEVICE_FATAL event has occurred in order to change its state
accordingly.

Extends the async_event function to get the context in all providers and
use it in mlx5 driver. Downstream patches from this series will use this
fatal state indication.

Change-Id: I762d690d3030e7a315bdcd3c5a7db1d64bdb88d6
Signed-off-by: Erez Shitrit <erezsh@mellanox.com>
Reviewed-by: Alex Vesker <valex@mellanox.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>