mlx5: Expose steering domain functionality
Domain is the frame for all of the dr (direct rule) objects.
There are ifferent domain types which also affect the object under that
domain. Each domain can hold multiple tables which can hold multiple
matchers and so on, this means that all of the dr (direct rule) objects
exist under a specific domain.  The domain object also holds the
resources needed for other objects such as memory management and send
engine.

Change-Id: I0e6b10403e90ff0dab6895b6670e6a3d82e07d0f
Signed-off-by: Alex Vesker <valex@mellanox.com>
Reviewed-by: Erez Shitrit <erezsh@mellanox.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>