mlx5: Meter action support
Meter object implemented by FW, the meter object points
to a given SW owned flow table (Policer).
The table is provided to the object by table_type and table_id.

The meter object sets the color of the traffic before the
policer based on a user given values during creation.

To pass traffic to the meter we set the hit address to the
meter ICM address rx/tx. this address is obtained on meter
action creation by using a query object command.

Note: This is a temporary implementation.

Issue: 1349276
Change-Id: If3252a178ef11dae704fc8b7666b3667eb99c4cd
Signed-off-by: Alex Vesker <valex@mellanox.com>